refactor: 에러 핸들링 정리 및 파일 분리
